+++ ndonnx/_core/_numericimpl.py
@@ -739,7 +739,7 @@ class NumericOperationsImpl(UniformShapeOperations):
             and isinstance(x.dtype, dtypes.Numerical)
         ):
             x, min, max = promote(x, min, max)
-            if isinstance(x.dtype, dtypes._NullableCore):
+            if isinstance(x.dtype, dtypes.NullableCore):
                 out_null = x.null
                 x_values = x.values._core()
                 clipped = from_corearray(opx.clip(x_values, min._core(), max._core()))
@@ -856,7 +856,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def all(self, x, *, axis=None, keepdims: bool = False):
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             x = ndx.where(x.null, True, x.values)
         if functools.reduce(operator.mul, x._static_shape, 1) == 0:
             return ndx.asarray(True, dtype=ndx.bool)
@@ -866,7 +866,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def any(self, x, *, axis=None, keepdims: bool = False):
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             x = ndx.where(x.null, False, x.values)
         if functools.reduce(operator.mul, x._static_shape, 1) == 0:
             return ndx.asarray(False, dtype=ndx.bool)
@@ -898,7 +898,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def tril(self, x, k=0) -> ndx.Array:
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             # NumPy appears to just ignore the mask so we do the same
             x = x.values
         return x._transmute(
@@ -909,7 +909,7 @@ class NumericOperationsImpl(UniformShapeOperations):
 
     @validate_core
     def triu(self, x, k=0) -> ndx.Array:
-        if isinstance(x.dtype, dtypes._NullableCore):
+        if isinstance(x.dtype, dtypes.NullableCore):
             # NumPy appears to just ignore the mask so we do the same
             x = x.values
         return x._transmute(